requirements.txt
transformers
openai
pandas
torch 
torchvision 
torchaudio
numpy
matplotlib
scikit-learn
Pillow
chromadb
streamlit